Remove dependency cycle on vendoring infra.
